requirements.txt
Django==2.1.9
django-appconf==1.0.3
django-cors-headers==3.1.1
djangorestframework==3.9.4
